Explicitly halt when returning the response
Previously the `POST /files` endpoint returned the URL of the resource
in the response body, because the result of `#created!` was an
assignment of the "Location" header, which returns the assigned value,
and since it's a String, Roda took it as the response body.

To prevent this, we explicitly say that we are returning the response,
so that whatever we assigned gets returned, and only that. This is also
the correct way to use Roda.