Always return 204 for OPTIONS requests
An OPTIONS is a "preflight" request part of the CORS specification, and
it should return a successful response if the request is safe to send.

Previously tus-ruby-server would returning a 404 on an `OPTIONS
/files/:uid` request if the file wouldn't exist. However, that wasn't
correct behaviour, because browser identifies this as a "preflight"
request failed, and returns a 500.

That means that if the browser would make an OPTIONS request for
resuming uploading a file that is still cached in local storage, but is
deleted due to past expiration date, tus-js-client (and likely other tus
client libraries) would consider this a failed request, instead of
uploading the file again from scratch.

Fixes #2