Don't return 404 when deleting a non-existing file
If the expiration time is set for a shorter time, it can likely happen
that the client tries to delete a file that has already expired.
However, tus-ruby-server currently returns a 404, because the file
doesn't exist.

I think this isn't correct behaviour, because requesting to DELETE a
file that doesn't exist could happen for various reasons, so returning a
404 in my opinion adds unnecessary brittleness. There is nothing invalid
about trying to delete a file that doesn't exist (and that makes sense
from the RESTful endpoints perspective in general).