Don't depend on Rack 2.x, support 1.x as well
Rails versions older than 5.x depend Rack "~> 1.6", and tus-ruby-server
currently requires "~> 2.0". This means that older versions of Rails
aren't currently able to include tus-ruby-server in the same Gemfile.

The only reason for depending on Rack 2.x was that Rack::File has a
different interface. However, we can easily make a conditional to
support Rack 1.x as well. This change is worthwhile, since many people
are likely using a web framework that is locked to Rack 1.x.

Fixes #1