Refresh Upload-Expires on PATCH requests
Previously it wouldn't be safe to set shorter expiration times, because
it would be set once in the beginning of upload, so you would need to
account for how long the upload is going to take.

However, according to the tus specification, the server can change the
expiration date over time. This means that we can refresh it on each
PATCH request, which allows the user to set a shorter expiration time,
to delete abandoned files that are not being actively uploaded much
quicker.

Since I don't see any reason why this shouldn't be the default
behaviour, this commit modifies the server to automatically update
expiration dates on each PATCH request.

Fixes #3