Don't read body for validating chunk length
Most web servers load the request body into an in-memory StringIO object
when it's small, but when it's large they use a Tempfile (Puma) or wrap
the TCP socket into an IO object for reading (Unicorn), to save on
memory.

That means if we just read the whole request body in memory, we're
essentially cancelling this memory optimization. Instead of doing that,
we can make sure that the request body is only read in chunks.

In the particular case of validating the chunk length, we don't even
need to read the request body, we can just read the "Content-Length"
header.