Read body in chunks for calculating checksum
Most web servers load the request body into an in-memory StringIO object
when it's small, but when it's large they use a Tempfile (Puma) or wrap
the TCP socket into an IO object for reading (Unicorn), to save on
memory.

That means if we just read the whole request body in memory, we're
essentially cancelling this memory optimization. Instead of doing that,
we can make sure that the request body is only read in chunks.

All Ruby digest implementations support incrementally calculating the
result chunk by chunk, so we can change Tus::Checksum to accept the
request body "rewindable input".