Save request body to storage in streaming fashion
Most web servers load the request body into an in-memory StringIO object
when it's small, but when it's large they use a Tempfile (Puma) or wrap
the TCP socket into an IO object for reading (Unicorn), to save on
memory.

That means if we just read the whole request body in memory, we're
essentially cancelling this memory optimization. Instead of doing that,
we can make sure that the request body is only read in chunks.

This commit updates the storages to accept an IO object (request body),
and save the content in a streaming fashion, loading only 16 KB of the
body at a time. This is unfortunately a backwards incompatible change,
any external storages will have to update their #upload method to
support an IO object.