Update :length and :uploadDate in Gridfs storage
It's useful to be able to know how large the uploaded files are, and
:length doesn't update itself. It's also useful to know when the file
has finished uploading, we will later use that when expiring files.

This brings backwards incompatibility where the sent chunk size has to
be evenly dividable by the Gridfs chunk size, otherwise some chunks will
be smaller than the specified chunk size, and then
`Gridfs#download_file` will throw an error saying that it encountered
chunk which was smaller than the chunk size but that it wasn't the last
chunk.