Stream file content on GET request
Currently the abstraction for retrieving files is that the storage needs
to return a path to a file on the disk. This works for filesystem
storage, since it can just return path to the file which is already
saved, however, this means that other storages like Gridfs and later S3
need to first download the whole file to disk, before they can start
streaming it to the response body. That's not very fast.

So we change the abstraction of storages from `#download_file` to
`#get_file`, which returns a "response" object that responds to `#each`
and `#close`, which can theoretically be returned directly as a Rack
response. The `#each` method will lazily retrieve file content in
chunks, so this way we can stream the file directly to the response
body, without any intermediate step. We use Roda's streaming plugin to
provide an abstraction for that, which is better than returning the
response directly because it also has EventMachine integration.

We also take the extra effort to support "Range" queries, so that files
like videos and audios can be served directly from the tus storage
(while they're being uploaded to a permanent storage) via an video/audio
player, which execute "Range" queries under-the-hood.