Fix GridFS not working correctly with Tus::Input
Previously Tus::Input would only be able to tell whether it has reached
EOF by attempting to read the next chunk, and realizing that it is nil.
However, this means that if we read chunks until EOF, we will get an
additional nil-chunk at the end, which breaks GridFS because it assumes
that every chunk is a String.

We fix this by checking whether we have reached EOF by comparing how
many bytes we have read vs what's the total size of the input. The
`#size` method isn't strictly part of the Rack specification in the list
of methods that the input object should respond to, but all major web
servers use input objects that respond to `#size`:

* Puma (StringIO, Tempfile)
* Unicorn (Unicorn::TeeInput)
* Passenger (PhusionPassenger::Utils::TeeInput)
* Thin (StringIO, Tempfile)
* WEBrick (StringIO)
* Other handlers that ship with Rack (Rack::RewindableInput)

It would be easier if the input object given by the web server simply
responded to `#eof?` (then we could just delegate to it), but in Unicorn
`#eof?` is a private method, and Passenger doesn't implement it.

Fixes #4