Have storages raise a "not found" error
Currently we're calling `#file_exists?` before each HEAD/GET/PATCH
request, which is cheap for FileSystem and Gridfs, but is expensive for
storages like S3 (which I intend to add soon), because it's an
additional HTTP request.

Instead we require storages to raise a Tus::NotFound error themselves if
the file is missing. Now S3 can attempt to fetch the info for a given
uid, and raise an error if fetching failed; that way only a single HTTP
request will be issued regardless of whether the file existed or not.

We still need to keep the `#file_exists?` method for Tus::Expirator, but
we will soon rewrite Tus::Expirator not to require this method.