Delegate expiration to storages and require manual triggering
Previously we were the ones implementing expiration in Tus::Expirator
using available storage methods. However, this implementation would be
very difficult to reuse for the future S3 storage, and also it wouldn't
be very efficient. So, in order to accomodate for S3, we delegate file
expiration to the storage directly, which allows the storage to
implement it in the most optimal way.

We also remove automatic file expiration, as the current implementation
can be very error prone, and errors wouldn't bubble up anywhere (for
Ruby < 2.4 the errors wouldn't even be logged). Also, the behaviour of
only triggering expiration when there are new requests is really ugly
design. It's much better and more straightforward to require users to
setup a recurring task that would call expiration directly.