Correctly generate and compare checksums
The Digest standard library offers three forms of digests:

  Digest::MD5.digest("...")
  Digest::MD5.hexdigest("...")
  Digest::MD5.base64digest("...")

The latter two are just convenience methods, they can both be calculated
from the first one:

  Digest::MD5.hexdigest("...") == Digest::MD5.digest("...").unpack("H*")[0]
  Digest::MD5.base64digest("...") == Base64.encode64(Digest::MD5.digest("..."))

When the tus client sends a base64-encoded checksum, we are currently
comparing the checksum in the wrong way; we were base64-decoding the
received checksum and comparing it to `#hexdigest`, while we should
actually be comparing it to raw `#digest`.

Since the upcoming S3 storage will require MD5 checksums to be
base64-encoded, we modify Tus::Checksum to generate *proper*
base64-encoded checksums by default.