Add #length method to result from #get_file
It's useful to be able to be able to retrieve the content length
directly from the response object, especially when you want to interact
with uploaded objects from outside the tus server.

In shine-tus-demo I want to provide an example of how to download the
file directly from the tus storage, using the Down::ChunkedIO wrapper.
Currently to get the content length we would have to make two HTTP
requests and leak details about the tus specification:

  storage = Tus::Server.opts[:storage]
  response = storage.get_file(uid)
  info = storage.get_info(uid)
  length = info["Upload-Length"].to_i

  Down::ChunkedIO.new(
    chunks: response.each,
    size:   length,
    close:  ->{response.close},
  )

With this commit we can read the length directly from the response
object:

  response = Tus::Server.opts[:storage].get_file(uid)

  Down::ChunkedIO.new(
    chunks: response.each,
    size:   response.length,
    close:  ->{response.close},
  )