Make :chunkSize static again, and allow input to be any size
There is no real reason to infer :chunkSize from the first uploaded
chunk, and it complicates validation for concatenation. So we make the
:chunkSize static again, 256KB by default.

We also make changes that don't require uploaded chunk length to be
multiple of :chunkSize anymore. We do that by detecting on each
`#patch_file` call whether the last Gridfs chunk is incomplete, and if
it is we fill it with the beginning of the input, then continue
splitting the rest of the input into new chunks.