Lazily require libraries for calculating checksums
That way only standard libraries that are used will be required. For
example, before this change, "base64", "digest", and "zlib" standard
libraries would all be required even if the developer isn't using the
checksum feature at all. If the developer is generating
SHA{1,256,384,512} or MD5 signature they don't need to require "zlib" or
"base64", and if they're using CRC32 they don't need to require
"digest".