Allow "Transfer-Encoding: chunked" requests
When a request is "Transfer-Encoding: chunked", it means
"Content-Length" is blank. We previously relied on "Content-Length"
being present, but we don't have to; this PR changes the server to rely
only on the request body length.

Ideally we would want to stream the request body to the storage, but
Amazon S3 doesn't support "Transfer-Encoding: chunked" (that is, it
requires that we specify "Content-Length"), so it's not really worth it.