Make only one thread raise in S3#concatenate
Currently all threads will propagate the exception to the main thread,
because Thread#abort_on_exception is set to true. However, if one thread
raises an exception, and the exception starts being rescued, then
another thread raises an exception, I'm not really clear what will
happen.

It's safer that only the first error is ever raised, and we do that by
leaving Thread#abort_on_exception to be false, and using the fact that
Thread#value will detect whether the thread raised an exception and
propagate it in the main thread.