Terminate S3 connection when response is closed
We don't have an explicit way to close the HTTP connection before time,
so we left Response#close unimplemented. However, AWS SDK will close the
HTTP connection if block given to Aws::S3::Object#get exits (there is an
"ensure" block). So, with the help of Fibers, we implement #close to
break out of the chunks block.