Improve streaming uploads
Firstly, we stop relying on Rack input responding to #size. One reason
is that it's not part of the Rack specification, so not all Rack inputs
need to respond to #size (Unicorn::StreamInput). Another reason is that,
for inputs that dynamically read from the socket, calling #size eagerly
reads whole data from the socket (Unicorn::TeeInput,
Passenger::Utils::TeeInput).

This means that, if the user uses `Transfer-Encoding: chunked` for PATCH
requests, we don't know upfront what is the size of the input, so we
update Tus::Input to check whether read content has exceeded the limit.

We also improve streaming uploads in Tus::Storage::Gridfs and
Tus::Storage::S3. For Gridfs we limit the memory usage by batching
inserts, so even if the client sends PATCH requests larger than 5MB,
Gridfs will only load 5MB into memory at once. For S3 we use a similar
strategy for the case when Content-Length is not known, because S3
doesn't support uploads of unknown length.