Include CORS and tus response headers in 404s
The error_handler Roda plugin creates a new response object, so any
response headers previously written will not be included in the
response. However, we are using error_handler also for validation
errors, and we want CORS and tus response headers to be included in
those responses.

It seems that the error_handler plugin should indeed be used for
erroneous behaviours, so we localize exception rescuing to where it can
happen.