Final improvements to streaming uploads
First, we remove detection of when Rack input is already buffered
(StringIO or Tempfile), because it doesn't bring real performance
optimizations, so it brings unnecessary complexity.

We also rename Tus::Input#bytes_read to Tus::Input#pos, because it has
the exact same semantics as IO#pos, so we might as well give it the same
name.

We modify all storages to return the number of bytes uploaded, and this
is now a requirement for all storages. We do that because the S3 storage
will not always be able to upload the chunk given to it, notably when
the chunk is smaller than 5MB and it's not the last chunk (since AWS S3
requires that all parts except the last one are 5MB or larger).

We also simplify Tus::Storage::S3#patch_file to always upload in chunks,
because (a) the performance benefits are not worth the complexity, and
(b) tus-js-client always sends the "Content-Length" header, even when it
tries the upload the whole file in a single request, so in that case we
should still upload as much as we can (Tus::Input#size was previously
returning value of the "Content-Length" header).

Finally, we improve performance of Tus::Storage::S3#patch_file by
buffering chunks in memory instead of on disk. That way we're doing less
IO for the cost of higher memory usage, which is in my opinion a
completely acceptable tradeoff because we won't be loading more than
10MB into memory at once. That way users can completely skip the disk if
they choose a Rack input that doesn't cache data on disk.