Add Goliath integration
Goliath is the optimal web server to run tus-ruby-server on, because by
utilizing EventMachine it's asynchronous both in reading the request
body and writing to the response body, so it's not affected by slow
clients.

It also allows tus-ruby-server to handle interrupted requests, by saving
data that has been received from the client until the interruption. This
means the ":chunkSize" tus-js-client option is not mandatory to achieve
resumable uploads, so the client can choose to use smaller number of
requests in situations of high network latency.