Resolve FiberError with S3 storage
Web servers like Goliath will stream the response in a separate thread
than the main worker thread. This means that if we initialize the S3
response, and that response already starts the fiber, resuming that
fiber in another thread will raise an FiberError exception, because Ruby
minds when you try to manipulate a fiber across threads.

We fix this by actually starting the fiber only when #each is called,
which should ensure it runs whole inside a single thread.