Improve handling aws-sdk fallback in S3 storage
Previously if neither aws-sdk-s3 nor aws-sdk were present, a deprecation
warning for "aws-sdk 2.x" would get printed. We change this behaviour to
first attempt to require aws-sdk after not succeeding with aws-sdk-s3,
and only then print a deprecation warning.

Another undesired behaviour was that if neither aws-sdk-s3 nor aws-sdk
were present, user would get a LoadError saying that aws-sdk was failed
to get required. This isn't what we want, because the user should
actually add aws-sdk-s3, not aws-sdk. We fix that by re-raising the
original aws-sdk-s3 LoadError if aws-sdk failed to get required.