Don't use the Roda streaming plugin
In Roda 3 the streaming plugin is really more like a nice wrapper for
when we are dynamically generating content. But here we already have a
Tus::Response object which is meant to be returned as the Rack response
body. So we don't need the extra complexity of the streaming plugin.