Add back support for Roda 2.x
Someone might be using tus-ruby-server with Roda 2.x, and they may not
wish to upgrade to Roda 3 just to get the latest changes.

There is no need to stop supporting Roda 2.x, we can just lock the
version which adds the feature of String/Integer class matchers as the
lowest version.

The reason we switched to this matchers in the first place is because on
Roda 3.x using "colon" identifiers in strings doesn't work anymore, at
least not by default.