Handle "name" and "type" metadata for Uppy compatibility
Uppy sends "name" and "type" metadata instead of "filename" and
"content_type", so we handle those in the download endpoint.

We also adopt is as the convention, as it's shorter and less coupled to
the header names. The "filename" and "content_type" metadata will
probably continue to be supported indefinitely.