Reduce memory in S3 storage and remove threading
We read chunks of data from the Rack input, and once those chunks are
uploaded, they can be deallocated to reduce memory pressure.

While here, we also remove threading from the upload. It was originally
added for reducing the time spent in the event loop when using
EventMachine. However, I think it's a non-problem and not worth the
extra complexity; any issues with request throughput should be solved by
increasing the number of EventMachine instances.