Allow using Rack::Sendfile to serve files saved to disk through the application server
Serving files through the Ruby application is very inefficient, because
the web worker is blocked until the user downloads the file, which can
significantly impact request throughput and make the application
vulnerable to DoS attacks.

The Rack::Sendfile middleware enables delegating file serving to the
underlying application server (Nginx, Apache). For it to work, the
response body object needs to respond to #to_path, so we extend the
result of FileSystem#get_file with that method.