Recommend using Falcon web server
Goliath built on top of the EventMachine ecosystem, which is dying for a
long time now. And Goliath, even though it's a web server and attempts to
bring Rack compatibility to EventMachine, doesn't have the ability to
act as a web server for Rack apps. Therefore I had to create
goliath-rack_proxy as an additional layer which provides that
functionality, and I wasn't happy with how difficult that was.

Falcon, on the other hand, is a web server built on top of the rich
"async" ecosystem. Like EventMachine, "async" also brings non-blocking IO
to Ruby, but instead of implementing the whole reactor in C++ like
EventMachine does, async has a much more layered implementation.

At the core of async is the nio4r gem, which implements a non-blocking
IO selector in C as native extensions. The async gem then implements the
reactor pattern using Nio4r and Ruby fibers. Fibers enable users of
async to avoid writing asynchronous code using callbacks, and instead
write it in a synchronous way.

The async-http gem then adds the HTTP layer, providing both an HTTP
server and an HTTP client. It uses async-io, which provides wrappers for
ruby `IO` subclasses that automatically hook up to async's reactor.

Finally, Falcon implements web server functionality on top of
async-http, and comes with a Rack adapter. This makes Falcon a
well-designed web server built with nice layers of abstraction.