Avoid using exceptions to break out when Queue is empty by closing it
Fetching items from the queue will block if queue is empty, to allow
queue to get new items in.

  queue = Queue.new
  queue.pop # will wait indefinitely

However, since we're already populating the queue with all the items
that will be in the queue upfront, we can close the queue, and that will
prevent fetching items to block.

  queue = Queue.new
  queue.close
  queue.pop #=> nil