Don't default Content-Type to application/octet-stream in download endpoint
When the file type was not specified in metadata, we default the
"Content-Type" header value to "application/octet-stream". This means
the browser will always treat it as generic binary content, and won't
attempt to determine the type from response body.

However, browser's automatic type detection can be useful, as it means
some things will work even when type wasn't explicitly set when the
upload was created. I also cannot see a scenario where it's more useful
to return "application/octet-stream" instead of nothing.