Simplify concatentation implementation
tusd doesn't parallelize validating partial uploads and apparently
that's good enough for them, so that's good enough for us. This
simplifies the validation logic significantly.

Since we have to fetch info objects of each partial upload anyway for
validation, we move calculating total length out of storages.

We additionally validate that partial uploads are completed.