Adjust multipart part size to AWS S3 limits
If the user is uploading a file larger than 50 GB, the minimum part size
of 5 MB is not enough anymore, because AWS S3 allows maximum of 10,000
multipart parts. So we adjust the multipart part size according to the
upload length, using the same logic as tusd uses.

We also reject files larger than 5 TB, as that's the maximum object size
AWS S3 accepts.

Finally, we make the limits configurable, in case someone is using an
S3-compatible service with different limits.

Fixes #14